pyproject.toml
[project]
name = "multi-agent-research"
version = "0.1.0"
description = "Multi-agent research assistant built with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.10.0",            # async HTTP for Tavily search
    "aiosqlite>=0.22.1",          # async SQLite for future use
    "beautifulsoup4>=4.14.3",
    "hatchling>=1.29.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langgraph>=1.0.10",
    "psycopg>=3.3.3",             # psycopg3 for Supabase PostgreSQL
    "pydantic>=2.12.5",
    "python-docx>=1.1.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.22",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
